+++ src/app/(site)/[slug]/page.tsx
@@ -5,10 +5,7 @@ import gallery from '@/data/gallery.json';
 export default async function ServicePage() {
   return (
     <>
-      <Gallery
-        title={gallery.titleServicePage}
-        labelledbyId="gallery-section-labelledby"
-      />
+      <Gallery title={gallery.titleServicePage} />
     </>
   );
 }
