+++ src/app/(site)/page.tsx
@@ -1,6 +1,6 @@
-import { CardsKreatywny } from '@/components/base';
-import { SliderTest } from '@/components/base/SliderTest';
-import { ContactLinks } from '@/components/ui';
+// import { CardsKreatywny } from '@/components/base';
+// import { SliderTest } from '@/components/base/SliderTest';
+// import { ContactLinks } from '@/components/ui';
 
 import { Gallery } from '@/sections';
 
@@ -9,26 +9,23 @@ import gallery from '@/data/gallery.json';
 export default async function Home() {
   return (
     <>
-      <Gallery
-        title={gallery.titleHomePage}
-        labelledbyId="gallery-section-labelledby"
-      />
+      <Gallery title={gallery.titleHomePage} />
 
-      <div className="container mt-10">
+      {/* <div className="container mt-10">
         <ContactLinks
           location={'header'}
-          className={'text-xl tracking-[-0.2px] smOnly:hidden'}
+          className={'smOnly:hidden text-xl tracking-[-0.2px]'}
         />
-      </div>
+      </div> */}
 
-      <div className="container bg-color-bg-primary">
+      {/* <div className="bg-color-bg-primary container">
         <CardsKreatywny />
         <ContactLinks location={'contacts'} className={'xl:items-start'} />
-      </div>
+      </div> */}
 
-      <div className="container my-[50px]">
+      {/* <div className="container my-[50px]">
         <SliderTest />
-      </div>
+      </div> */}
     </>
   );
 }
