+++ src/components/base/index.ts
@@ -1,1 +1,2 @@
 export * from './CardsKreatywny';
+export * from './GalleryView';
