+++ src/components/ui/SectionTitle/SectionTitle.tsx
@@ -2,7 +2,6 @@ import { SectionTitleProps } from './type';
 
 export const SectionTitle: React.FC<SectionTitleProps> = ({
   children,
-  ariaLabelledbyId,
   accentColor,
   center,
   className = '',
@@ -14,7 +13,6 @@ export const SectionTitle: React.FC<SectionTitleProps> = ({
 
   return (
     <h2
-      id={ariaLabelledbyId}
       className={`${fontColor} ${isCenteredText} text-[36px] font-extrabold leading-[1.1] 
       tracking-[-0.36px] md:text-[48px] md:tracking-[-0.48px] xl:text-[56px]
       xl:tracking-[-0.56px] ${className}`}
