+++ src/components/ui/SectionTitle/type.ts
@@ -2,7 +2,6 @@ import { ReactNode } from 'react';
 
 export type SectionTitleProps = {
   children: ReactNode;
-  ariaLabelledbyId: string;
   accentColor?: true;
   center?: true;
   className?: string;
