+++ src/components/ui/SliderControls/SliderControls.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import PrevIcon from '~/icons/arrow-left.svg';
 import NextIcon from '~/icons/arrow-right.svg';
 
 import common from '@/data/common.json';
@@ -22,7 +21,12 @@ export const SliderControls: React.FC<SliderControlsProps> = ({ section }) => {
         className={`button-prev-${section} flex h-14 w-14 items-center justify-center rounded-full border border-solid border-color-ctrl-btn-border bg-color-secondary transition-all hover:border-transparent hover:bg-color-bg-primary focus:border-transparent focus:bg-color-bg-primary`}
         type="button"
       >
-        <PrevIcon aria-label={prevBtnArialabelIcon} width={24} height={24} />
+        <NextIcon
+          className="rotate-180"
+          aria-label={prevBtnArialabelIcon}
+          width={24}
+          height={24}
+        />
       </button>
 
       <button
