+++ src/components/ui/index.ts
@@ -3,6 +3,6 @@ export * from './Logo';
 export * from './CardKreatywny';
 export * from './ContactLinks';
 export * from './Slider';
-export * from './Title';
+export * from './SectionTitle';
 export * from './LightboxImage';
 export * from './GalleryCard';
