+++ src/sections/Gallery/Gallery.tsx
@@ -1,18 +1,13 @@
 import { SectionTitle } from '@/components/ui';
-import { GallerySectionProps } from './types';
-import { GalleryView } from '@/components/base/GalleryView';
+import { GalleryView } from '@/components/base';
 
-export const Gallery: React.FC<GallerySectionProps> = ({
-  title,
-  labelledbyId,
-}) => {
+import { GallerySectionProps } from './types';
+
+export const Gallery: React.FC<GallerySectionProps> = ({ title }) => {
   return (
-    <section
-      aria-labelledby={labelledbyId}
-      className="pt-12 md:pt-16 xl:pt-[124px]"
-    >
+    <section id="gallery" className="section">
       <div className="container flex flex-col gap-9 md:gap-12">
-        <SectionTitle ariaLabelledbyId={labelledbyId} accentColor center>
+        <SectionTitle accentColor center>
           {title}
         </SectionTitle>
