+++ src/sections/Gallery/types.ts
@@ -1,4 +1,3 @@
 export type GallerySectionProps = {
   title: string;
-  labelledbyId: string;
 };
